chore: extract two column layout into generic layout component
