+++ src/renderer/src/views/Migration/styles.ts
@@ -1,10 +1,4 @@
 import styled from '@emotion/styled'
-import { Column } from '@renderer/components/LayoutComponents'
-
-export const Contents = styled(Column)`
-  justify-content: center;
-  align-items: center;
-`
 
 export const BoldSpan = styled('span')`
   font-weight: 700;
